+++ src/hooks/index.ts
@@ -1,5 +1,6 @@
 import useAppDispatch from './useAppDispatch'
 import useAppSelector from './useAppSelector'
+import useClickOutside from './useClickOutside'
 import useExpandableGroups from './useExpandableGroups'
 import useInitConfig from './useInitConfig'
 import usePWAInstall from './usePWAInstall'
@@ -7,6 +8,7 @@ import usePWAInstall from './usePWAInstall'
 export {
   useAppDispatch,
   useAppSelector,
+  useClickOutside,
   useExpandableGroups,
   useInitConfig,
   usePWAInstall,