+++ src/pages/Dashboard/Transactions/index.tsx
@@ -2,14 +2,12 @@ import { useEffect, useState } from 'react'
 import { useOutletContext } from 'react-router-dom'
 import { Virtuoso } from 'react-virtuoso'
 import { useIntl } from 'react-intl'
+import type { LayoutContextType } from '@/components/Layout'
+import { DEFAULT_EXPANDED_COUNT } from '@/constants/config'
 import { useAppSelector, useExpandableGroups } from '@/hooks'
 import { formatMonthLabel, shouldShowMonthHeader } from '@/utils'
 import SearchResult from './SearchResult'
 import TransactionContainer from './TransactionContainer'
-import type { LayoutContextType } from '@/components/Layout'
-
-// Number of most-recent groups expanded by default on first load.
-const DEFAULT_EXPANDED_COUNT = 3
 
 const Transactions = () => {
   const { scrollParent } = useOutletContext<LayoutContextType>()