+++ src/store/report/report-actions.ts
@@ -1,5 +1,5 @@
 import { PayloadAction } from '@reduxjs/toolkit'
-import { Data, DashboardInfo } from '@/types'
+import { Data, DashboardInfo, Category } from '@/types'
 import { InitialState } from './report-slice'
 
 export const setDashboardInfo = (
@@ -18,8 +18,32 @@ export const setDashboardInfo = (
 
 export const setDetailData = (
   state: InitialState,
-  action: PayloadAction<Data[]>
+  action: PayloadAction<{
+    data: Data[]
+    detailCount: number
+    detailIncome: number
+    detailExpense: number
+    detailBudget: number
+    detailRemainingBudget: number
+    selectedDetailCategory: Category | null
+  }>
 ) => {
-  state.detailData = action.payload
+  const {
+    data,
+    detailCount,
+    detailIncome,
+    detailExpense,
+    detailBudget,
+    detailRemainingBudget,
+    selectedDetailCategory,
+  } = action.payload
+
+  state.detailData = data
+  state.detailCount = detailCount
+  state.detailIncome = detailIncome
+  state.detailExpense = detailExpense
+  state.detailBudget = detailBudget
+  state.detailRemainingBudget = detailRemainingBudget
+  state.selectedDetailCategory = selectedDetailCategory || null
   state.isDetailLoading = false
 }