+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 
 All notable changes to this project will be documented in this file. See [commit-and-tag-version](https://github.com/absolute-version/commit-and-tag-version) for commit guidelines.
 
+## [1.0.0-beta.22](https://github.com/troy213/expense-tracker-v2/compare/v1.0.0-beta.21...v1.0.0-beta.22) (2026-05-24)
+
+
+### Features
+
+* update report detail page UI ([967d55e](https://github.com/troy213/expense-tracker-v2/commit/967d55ea118a145c3f26dc566c7dd742f84b1fd7))
+* update report detail UI ([24e7a60](https://github.com/troy213/expense-tracker-v2/commit/24e7a60a856322d61b003e0e6a3db621d3f6a32b))
+
+
+### Bug Fixes
+
+* outside click form select and date range modal issue ([f813400](https://github.com/troy213/expense-tracker-v2/commit/f81340078bc1835249e5264739c432df465f70f2))
+
 ## [1.0.0-beta.21](https://github.com/troy213/expense-tracker-v2/compare/v1.0.0-beta.20...v1.0.0-beta.21) (2026-05-23)